ms_clusterizacao/requirements.txt
# Databricks
databricks.sql

# Data Science
pandas
numpy

# Machine Learning
scikit-learn